Refactors workflows to module and starts paradag support
